requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database - Async Support
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
msal==1.31.0

# Configuration & Environment
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Document Processing
python-pptx==1.0.2
PyPDF2==3.0.1
openpyxl==3.1.5

# NLP & Text Analysis
spacy==3.8.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
vaderSentiment==3.3.2

# LLM Integration
openai==1.51.2
anthropic==0.39.0

# Utilities
requests==2.32.3
aiofiles==24.1.0
numpy>=2.1.3

# Development & Testing (optional)
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
